Stop state machine before destroying it, to ensure that client read streams are not running.
Summary:
If the ReplicatedStateMachine gets destroyed while its client read streams are still registered, their callbacks point to freed memory. This causes crash due to use-after-free if a gap or record arrives later.
To prevent that, the state machine now calls stop() in the destructor to destroy client read streams.

The crash may happen for instance if the state machine receives a snapshot that has a read pointer embedded, which covers all of the remaining deltas. In that case the replicated state machine pretends that it reached the tail of the delta log, although it's not true. if the caller destroys the state machine at this point without calling stop() (which is the case in EventLogUtils), the read stream will keep reading and hit the bug.

Reviewed By: MohamedBassem

Differential Revision: D18695387

fbshipit-source-id: 16a666dab4a40981b0cc90b3adec694a968e7c8c