Remove node count validation from LocalLogsConfig
Summary: I understand the rationale behind this validation, but it's just weird because it's coupling the validity of the logs config with the nodes config. Also, in the NCM world, this really won't work, and it's not worth the hassle of special casing it here, so I'm just removing this check. This was preventing logdevice from running with an empty nodes section.

Reviewed By: AhmedSoliman

Differential Revision: D18706352

fbshipit-source-id: 5de1a84dc04729c9ac72ec7835cf8264f4e12dac