Reduce number of copies from socket if hit with ENOBUFS
Summary:
When worker is overloaded we avoid reading from the evbuffer. To push back, we
left the bytes in evbuffer and read it out everytime from the socket. This
works if the number of sockets per worker and size of payload is small
otherwise we end up spending a significant portion of the cpu on just copying
from socket and do not do any useful work on the cpu.

I am fixing this by copying the message payload just once and cache the buffer.
If we hit with nobufs the cache is kept in the socket and we retry from the
same cache next time.

Other option discussed was to try to get the token before copying out data and
fail fast. To do that I would have to change new and old socket interfaces and
some of the ownership like for example deps did not exist in ProtocolHandler ,
I would have to rethink how to get token hence I dropped the idea. But if you
guys think its better I am ok with dropping this diff.

This diff reintroduces removing data from the evbuffer, I am being little bold
here but otherwise the changes looked bad hence just removing it made sense. if we hit NOBUFS,
this time we save in cache which should be ok but please make sure it is the
case.

This diff also changes how available bytes are calculated when we enter
onBytesAvailable. Please make sure that is correct.

Because of this cached buffer , there are two places where we can have valid
data , evbuffer and this cache. Please make sure if I have fixed everywhere
necessary correctly.

This problem did not exist with new socket as we copied only once in that
scenario. Please verify the nobufs implementation for new socket looks good.

Reviewed By: tau0

Differential Revision: D18513530

fbshipit-source-id: 37f1065dde5ec3a5666ef0afb7c992912fe67ff7