Self Registration: Adding integration tests for self registration
Summary: This diff adds integration tests for node self registration. Self registered node are special because during their addition, we don't know their node index in advance which a lot of the integration test tooling depend on, that's why they are special cased in the integration test utils and don't use the normal expand utilities.

Reviewed By: AhmedSoliman

Differential Revision: D18706440

fbshipit-source-id: feca3479cf56817987372c11867c3d0b7156f7e1