DocumentationInSyncTest: Print a much nicer diff when a settings was added or removed
Summary:
To diff two text files, it used to compare i-th line of one file to i-th line of the other, and print all that don't match. If a line was added (as happens when you add a setting), all lines would shift, and the test would print lots of lines - ~half of the whole settings.md on average. The test also prints some useful notes after the diff, so if the diff is too long, the test runner would truncate the output and make the notes hard to find.

This diff moves the notes to the top, makes the diff logic handle added/removed lines well, and limits diff output to ~45 lines.

Reviewed By: gdrane

Differential Revision: D18925266

fbshipit-source-id: 8024572ddcbc4288421b9a4160201d3ca4fd7015