Add node version
Summary:
# What
This changeset adds a `version` field to the service discovery configuration.
The purpose of a `version` field is to allow a smart conflict resolution in some edge case when two nodes with the same name are fighting for the cluster membership.

Semantically, if two nodes have the same name, then the node with a higher version wins and preempts another one. All the mutable attributes in the corresponding part of nodes configuration will be updated with the attributes of a node with a higher version.

This change is backwards compatible.

# Why
This functionality might be useful in some edge cases, e.g. when running in a containerised environment with a network split.

Reviewed By: MohamedBassem

Differential Revision: D18448755

fbshipit-source-id: 3fce5a6de6ba85e868d924749b86c581bc4da090