Add NodeHealthStats to HealthMonitor
Summary:
A stats object is updated in `HealthMonitor::processReports() `and then copied into the  `health_stats_`  field to ensure that the  `health_stats_`  field contains coherent stats at all times.
In other words, this is to make sure that when accessed from the outside, it always contains stats from the same HM loop, and does not have half of it's values updated, and the other half belonging to an old loop.

Reviewed By: mcrnic

Differential Revision: D18708462

fbshipit-source-id: d4240832a85b23540e78ff15eacc808a4a42a343