Keep a pointer to StatsHolder in LogStorageStateMap
Summary:
During server startup, `LogStorageStateMap` gets initialized before the processor initialization. Since the processor also holds a reference to the `StatsHolder` used by `LogStorageStateMap`, any stat updates done before initializing the processor are lost. To avoid it, this patch stores `StatsHolder` pointer in `LogStorageStateMap` in addition to the `Processor` and uses it directly.

Added a test case to execute an error code path and also to verify that the stats are updated correctly even before the processor initialization.

Reviewed By: al13n321

Differential Revision: D19320724

fbshipit-source-id: f5eef6edf98f8bcd4bb997e18f4c08c17e22ef9d