ClientReadStream: update Sender state's next lsn on sendStart() to avoid bogus error messages
Summary:
When we rewind, we send START messages for a window starting at
next_lsn_to_deliver. However, if a storage node has not sent records/gaps up to
that lsn, we might get some log lines complaining about getting gaps with
advanced lsn (example: if N1 has sent data up to lsn 10 but our read pointer is
at lsn 20 and we rewind, we will issue a START message from lsn 20, thus
possibly getting a gap that starts at lsn 20, causing CRS to complain a bit).

This updates the next lsn for a storage node when we issue a START so no
error messages are issued in this situation.

Reviewed By: tau0

Differential Revision: D19309581

fbshipit-source-id: 7479cc718ba56a1bf8c3fd4dab73f082aa61ea16