make prometheus-cpp a submodule of LogDevice
If ldprometheus plugin is in the main build, its external dependencies
should be handled the same way all other LogDevice external dependencies are.

Test plan:

Did a clean pull of LogDevice with this change into a new directory.
Followed the steps to rebuild LogDevice normally.
Verified that build completes and the Prometheus plugin loads fine.