Deprecate the include_log_config field from ServerConfig
Summary:
We have a functionality to fallback to a logsconfig store in a remote location if LCM fails for any reason. This fallback was stored in the ServerConfig as the `include_log_config` field. LCM has been stable in prod for a while now. For the past year, we've never disabled it. The techdebt of maintaining the fallback functionality is very high and introduces complexity in a lot of places as proved by this diff.

This diff, deprecates and removes the ability to include logsconfig from remote places.

This diff touches
- LogsConfigParser to ignore the `include_log_config` field from the config.
- Modifying TextConfigUpdater to not fetch this remote logsconfig anymore.
- Modifying the Configuration.h file to not include a callback for fetching the remote files.
- Remove the tests that tests for this feature.
- Removes the ability to parse gzipped configs as it was only used for logsconfig includes.
- Modify the InfoConfig ldquery to not include the included logsconfig metadata.
- Remove other references to this field in the c++ codebase.

Reviewed By: AhmedSoliman, Zepheus

Differential Revision: D19474389

fbshipit-source-id: f88efd46ccb6da96bffe9f6b1028097b81820515