Decouple checking and updating internal state in SSLFetcher
Summary:
Make requireContextUpdate a const method, and provide a updateState method that updates the internal state of SSLFetcher.
Aslo add a stat to keep track of the number of SSL context created.

This prevents unnecessary re-creation of the SSL context that happens on the second SSL connection, due to last_loaded_ not set.

Reviewed By: gdrane

Differential Revision: D19644733

fbshipit-source-id: 8eaca913a1ac63d8fd1d196f6cbbe98c153ecd12