Searching for maintenance by node_id should match sequencer, shards, or both
Summary:
While investigating an issue in a disaggregated cluster, I found that our remediations won't remove the fbar maintenance. Digging deeper it turned out that our logic for searching for maintenances in ldops by node_id (supposed to be a helper) will need to match maintenances where both sequencers and shards match the input. I've changed this to match our real expectations. This now matches either nodes or sequencers, or both.

This works while keeping the semantics of that we have to match maintenances, otherwise we will return empty generator.

Reviewed By: spuzirev

Differential Revision: D19813742

fbshipit-source-id: 4c4add92e90ac8de902b1fc8691fdc75fbad3459