(reverting backout of D19322704) [logdevice] Make worker id and type part of EventLogStateMachine ctor
Summary: Today the EventLogStateMachine is hardcoded to always run on WG0. This is done by posting the StartEventLogStateMachine to run on WG0 and attaching the EventLogStateMachine object to that worker.  RebuildingCoordinator which acts on updates from EventLog is also hardcoded to run on WG0.  This diff is first part in trying to make EventLogStateMachine run on any worker. This diff is mostly mechanical change to make EventLogStateMachine's ctor take workerid and worker type as input. When StartEventLogStateMachineRequest is executed, it will run on the worker thread assigned during EventLogStateMachine creation. RebuildingCoordinator will also make use of this information to run on the same thread. The ctor's default to WG0. So everything runs on the same worker as today. Next diff in stack adds logic to run them on background thread

Reviewed By: AhmedSoliman

Differential Revision: D19858069

fbshipit-source-id: 3a1e2aa908e6441834306c6f366e13e56a30c389