Nodes Configuration: Start from generation 2 in NCM world
Summary:
Generations in NCM world (and more importantly in self registration world), should start from generation 2. Why? Consider this scenario:
- A node was added with storage state PROVISIONING and generation 1.
- The node marked itself as PROVISIONED and eventually moves to READ_WRITE.
- The disk gets wiped.
- On startup, the node will find that its generation is 1, so it assumes that it's a new node and bypasses rebuilding. This is silent under replication.

So the fix is that for nodes that was added in a world where the NCM is the source of truth (either via admin API or self registration), to start with generation > 1. Which means that it will rely *only* on the storage state check.

Eventually, once NCM is the only source of truth, generations should get deprecated.

Notes:
- All the servers that will pick this change and has self registration enabled, will start updating there generations on startup to 2.
- Any admin API `updateNode` call will also update node generation to 2.

Reviewed By: AhmedSoliman

Differential Revision: D20191071

fbshipit-source-id: 7c85b425e587354c70eea2ba37c3383f524d3dc6