Introducing logs info command
Summary:
This deprecates the internal `tail-attributes/head-attributes` and introduces a new `logs info id=<>` command that is also available for open-source. The command currently prints the tail LSN, the last real record LSN, the trim point information, and whether the log has data or not.

This is a useful tool that quickly grabs different pieces of information about any log. Can also be extended in the future to include more information (sequencer, etc.)

Reviewed By: MohamedBassem

Differential Revision: D20308605

fbshipit-source-id: 5ab94b326f1793ff7a3ebddc679fb78c39e29698