Dynamically change run interval for metadata trimming
Summary:
## Goal
We need a way to update/disable trimming of metadata log before even trying to go to production. This change should not require restart or it may severely reduce our ability to quickly remediate any issues with trimming.
## Pitfalls
Instances of `Timer` class may be accessed from single Worker thread only but updates of settings happen in random thread. Even worse, destruction of Timer must happen in the same Worker thread and there is no guarantees for this since `Sequencer` is multi-threaded.
## Approach
I re-wrote trimmer to delegate all interactions with `Timer` to single Worker thread chosen at the creation time.
The tricky part here is destruction: since destruction has to be delegated as well it may lead `use-after-free` bugs. Image following scenario:
1. Sequencer is being destructed so we publish request(1) to Worker thread to stop the timer.
2. At the same time settings get updated so we publish request (2) to update trimming interval (timer is still alive) and this request is queued after destruction.
3. Request 1 is executed, trimmer is destroyed so when it is time to execute request (2) there is no trimmer object to check current state.
To address this issue I extracted all needed fields into separate struct called `State` which is being carried over to callback and may outlive the trimmer which created it. As result the callback is always able to check whether trimmer has been stopped or not and act accordingly.

Reviewed By: tau0

Differential Revision: D19859251

fbshipit-source-id: 5f799482344b5be8746b273deeaf1f4c176e2b2a