Make RebuildingCoordinatorTest work with rebuilding v2
Summary: RebuildingCoordinatorTest was testing both RebuildingCoordinator and ShardRebuildingV1 (it's from the times when they were the same thing). This diff removes the test cases that were mostly about ShardRebuildingV1, and changes the test cases that are mostly about RebuildingCoordinator to not use ShardRebuildingV1 at all. The next diff in the stack removes ShardRebuildingV1.

Reviewed By: tau0

Differential Revision: D20603589

fbshipit-source-id: f2be6a4b280432ae67b658b5398c8b19d715e867